improve /run handling, token handling with /run
make this act more reasonable. we drop the x-auth-token header when
using /run as the idea of that endpoint is to use whats in the lowstate.
we get rid of the /token route handling as it seemingly does the same
thing as /login, and doesnt actually exist in the rest_tornado impl so
I'm considering it deprecated for simplicity.